1.1_Multi_Agent_Centralized_A2A/Client/requirements.txt
a2a-sdk[http-server]
uvicorn
python-dotenv